requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
redis
bcrypt
passlib
jose
pyjwt
requests